refactor(guard): resolve recovery review findings
