+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
-  "maximum_collected_cases": 15383,
-  "maximum_test_source_lines": 331841,
+  "maximum_collected_cases": 15393,
+  "maximum_test_source_lines": 332112,
   "schema_version": 1
 }
