+++ src/codex_plugin_scanner/guard/runtime/compound_git_inspection.py
@@ -9,6 +9,7 @@ import subprocess
 from pathlib import Path
 from typing import Final
 
+from .compound_git_bounds import safe_bound_segment
 from .git_execution_safety import (
     git_config_routing_environment_is_clean,
     git_fetch_origin_has_execution_free_config,
@@ -24,7 +25,6 @@ _OBJECT_EXISTENCE_QUERY: Final = re.compile(
     r"[A-Za-z0-9][A-Za-z0-9._/-]{0,255}(?:\^|\^\{(?:blob|commit|object|tag|tree)\})?"
 )
 _REPOSITORY_PATH_COMPONENT: Final = re.compile(r"[A-Za-z0-9_][A-Za-z0-9_.-]{0,127}")
-_BOUND: Final = 1000
 
 
 def canonical_home_git_c_path(command_text: str) -> str | None:
@@ -68,7 +68,7 @@ def is_low_risk_compound_git_inspection(context: ShellExecutionContext) -> bool:
                 return False
             continue
         if command in {"head", "tail"}:
-            if not _safe_bound_segment(segment, previous=context.segments[index - 1]):
+            if not safe_bound_segment(segment, previous=context.segments[index - 1]):
                 return False
             continue
         return False
@@ -380,9 +380,7 @@ def _safe_diff_args(args: tuple[str, ...]) -> bool:
 
 
 def _safe_diff_pathspec(value: str) -> bool:
-    if _safe_repository_path(value):
-        return True
-    return value.startswith((":!", ":^")) and _safe_repository_path(value[2:])
+    return _safe_repository_path(value) or (value.startswith((":!", ":^")) and _safe_repository_path(value[2:]))
 
 
 def _safe_show_args(args: tuple[str, ...]) -> bool:
@@ -554,17 +552,6 @@ def _safe_echo_segment(segment: ShellExecutionSegment) -> bool:
         and segment.control_after == ("&&",)
         and all(token not in {"-e", "-E", "-n"} and not _dynamic(token) for token in segment.tokens[1:])
     )
-
-
-def _safe_bound_segment(segment: ShellExecutionSegment, *, previous: ShellExecutionSegment) -> bool:
-    if segment.control_before != ("|",) or len(segment.tokens) != 2:
-        return False
-    if not previous.tokens or previous.tokens[0] != "git" or previous.control_after != ("|",):
-        return False
-    count = segment.tokens[1]
-    if not count.startswith("-") or not count[1:].isdigit():
-        return False
-    return 1 <= int(count[1:]) <= _BOUND
 
 
 def _dynamic(value: str) -> bool:
