+++ src/codex_plugin_scanner/guard/store_approvals.py
@@ -7,7 +7,6 @@ import hashlib
 import json
 import re
 import sqlite3
-from collections.abc import Mapping, Sequence
 
 from .approval_resolution import approval_resolution_block_reason
 from .approval_scope_support import request_scope_contract_payload, supported_request_scopes
@@ -285,246 +284,10 @@ def add_approval_request(
     *,
     oauth_source: str = "default",
 ) -> str:
-    canonical_decision = canonical_approval_surfaces(
-        request.policy_action,
-        request.decision_v2_json,
-        request.action_envelope_json,
-        reject_contradiction=True,
-    )
-    _begin_immediate(connection)
-    normalized_oauth_source = oauth_source.strip().lower() or "default"
-    identity_key = _normalized_identity_key(request.launch_target)
-    action_identity, queue_group_id = approval_queue_identity_for_request(request)
-    watch_only_observation = _scanner_evidence_is_watch_only(request.scanner_evidence)
-    existing = connection.execute(
-        """
-        select request_id
-        from approval_requests
-        where queue_group_id = ?
-          and harness = ?
-          and oauth_source = ?
-          and status = 'pending'
-        order by last_seen_at desc, request_id desc
-        limit 1
-        """,
-        (queue_group_id, request.harness, normalized_oauth_source),
-    ).fetchone()
-    if existing is None:
-        existing = connection.execute(
-            """
-        select request_id
-        from approval_requests
-        where harness = ?
-          and oauth_source = ?
-          and artifact_id = ?
-          and workspace IS ?
-          and normalized_identity_key = ?
-          and queue_group_id IS NULL
-          and status = 'pending'
-        order by created_at desc
-        limit 1
-        """,
-            (
-                request.harness,
-                normalized_oauth_source,
-                request.artifact_id,
-                request.workspace,
-                identity_key,
-            ),
-        ).fetchone()
-    if existing is None:
-        existing = connection.execute(
-            """
-            select request_id
-            from approval_requests
-            where harness = ?
-              and oauth_source = ?
-              and artifact_id = ?
-              and workspace IS ?
-              and launch_target IS ?
-              and normalized_identity_key IS NULL
-              and queue_group_id IS NULL
-              and status = 'pending'
-            order by created_at desc
-            limit 1
-            """,
-            (
-                request.harness,
-                normalized_oauth_source,
-                request.artifact_id,
-                request.workspace,
-                request.launch_target,
-            ),
-        ).fetchone()
-    request_id = str(existing["request_id"]) if existing is not None else request.request_id
-    if existing is not None:
-        review_command = _rewrite_review_command(request.review_command, request_id)
-        approval_url = _rewrite_approval_url(request.approval_url, request_id)
-        connection.execute(
-            """
-            update approval_requests
-            set harness = ?, artifact_name = ?, artifact_type = ?, artifact_hash = ?, publisher = ?, policy_action = ?,
-                recommended_scope = ?, changed_fields_json = ?, source_scope = ?, config_path = ?, workspace = ?,
-                launch_target = ?, normalized_identity_key = ?, action_identity = ?, queue_group_id = ?,
-                dedupe_count = coalesce(dedupe_count, 1) + 1, last_seen_at = ?,
-                transport = ?, risk_summary = ?, risk_signals_json = ?,
-                artifact_label = ?, source_label = ?, trigger_summary = ?, why_now = ?, launch_summary = ?,
-                risk_headline = ?, action_envelope_json = ?, decision_v2_json = ?, fallback_cli_command = ?,
-                scanner_evidence_json = ?,
-                watch_only_observation = case
-                    when watch_only_observation = 1 and ? = 1 then 1
-                    else 0
-                end,
-                browser_intent_json = ?, review_command = ?, approval_url = ?,
-                raw_command_text = ?, guard_version = ?,
-                first_seen_guard_version = coalesce(first_seen_guard_version, ?),
-                last_seen_guard_version = ?
-            where request_id = ? and oauth_source = ?
-            """,
-            (
-                request.harness,
-                request.artifact_name,
-                request.artifact_type,
-                request.artifact_hash,
-                request.publisher,
-                canonical_decision.policy_action,
-                request.recommended_scope,
-                json.dumps(list(request.changed_fields)),
-                request.source_scope,
-                request.config_path,
-                request.workspace,
-                request.launch_target,
-                identity_key,
-                action_identity,
-                queue_group_id,
-                now,
-                request.transport,
-                request.risk_summary,
-                json.dumps(list(request.risk_signals)),
-                request.artifact_label,
-                request.source_label,
-                request.trigger_summary,
-                request.why_now,
-                request.launch_summary,
-                request.risk_headline,
-                (
-                    json.dumps(canonical_decision.action_envelope_json)
-                    if canonical_decision.action_envelope_json is not None
-                    else None
-                ),
-                json.dumps(canonical_decision.decision_v2_json),
-                (
-                    _rewrite_review_command(request.fallback_cli_command, request_id)
-                    if request.fallback_cli_command
-                    else None
-                ),
-                json.dumps(list(request.scanner_evidence), sort_keys=True),
-                int(watch_only_observation),
-                json.dumps(request.browser_intent, sort_keys=True) if request.browser_intent is not None else None,
-                review_command,
-                approval_url,
-                request.raw_command_text,
-                request.guard_version,
-                request.first_seen_guard_version or request.guard_version,
-                request.last_seen_guard_version or request.guard_version,
-                request_id,
-                normalized_oauth_source,
-            ),
-        )
-        return request_id
-    connection.execute(
-        """
-        insert into approval_requests (
-          request_id, harness, artifact_id, artifact_name, artifact_type, artifact_hash, publisher, policy_action,
-          recommended_scope, changed_fields_json, source_scope, oauth_source, config_path, workspace,
-          launch_target, normalized_identity_key, action_identity, queue_group_id, dedupe_count, last_seen_at,
-          transport, risk_summary,
-          risk_signals_json, artifact_label, source_label, trigger_summary, why_now, launch_summary, risk_headline,
-          action_envelope_json, decision_v2_json, fallback_cli_command, scanner_evidence_json, browser_intent_json,
-          review_command, approval_url, status, resolution_action, resolution_scope, reason, created_at, resolved_at,
-          raw_command_text, guard_version, first_seen_guard_version, last_seen_guard_version,
-          watch_only_observation
-        )
-        values (
-          ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,
-          ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,
-          ?, ?, ?, ?, ?, ?, ?
-            )
-        """,
-        (
-            request.request_id,
-            request.harness,
-            request.artifact_id,
-            request.artifact_name,
-            request.artifact_type,
-            request.artifact_hash,
-            request.publisher,
-            canonical_decision.policy_action,
-            request.recommended_scope,
-            json.dumps(list(request.changed_fields)),
-            request.source_scope,
-            normalized_oauth_source,
-            request.config_path,
-            request.workspace,
-            request.launch_target,
-            identity_key,
-            action_identity,
-            queue_group_id,
-            max(1, int(request.dedupe_count)),
-            request.last_seen_at or now,
-            request.transport,
-            request.risk_summary,
-            json.dumps(list(request.risk_signals)),
-            request.artifact_label,
-            request.source_label,
-            request.trigger_summary,
-            request.why_now,
-            request.launch_summary,
-            request.risk_headline,
-            (
-                json.dumps(canonical_decision.action_envelope_json)
-                if canonical_decision.action_envelope_json is not None
-                else None
-            ),
-            json.dumps(canonical_decision.decision_v2_json),
-            request.fallback_cli_command,
-            json.dumps(list(request.scanner_evidence), sort_keys=True),
-            json.dumps(request.browser_intent, sort_keys=True) if request.browser_intent is not None else None,
-            request.review_command,
-            request.approval_url,
-            "pending",
-            None,
-            None,
-            None,
-            now,
-            None,
-            request.raw_command_text,
-            request.guard_version,
-            request.first_seen_guard_version or request.guard_version,
-            request.last_seen_guard_version or request.guard_version,
-            int(watch_only_observation),
-        ),
-    )
-    return request.request_id
+    """Compatibility facade for the focused approval write service."""
+    from .store_approval_writes import add_approval_request as persist_approval_request
 
-
-def _scanner_evidence_is_watch_only(scanner_evidence: Sequence[Mapping[str, object]]) -> bool:
-    return any(item.get("source") == "observe_mode_inbox" for item in scanner_evidence)
-
-
-def _rewrite_review_command(command: str, request_id: str) -> str:
-    prefix, _, _ = command.rpartition(" ")
-    if prefix:
-        return f"{prefix} {request_id}"
-    return request_id
-
-
-def _rewrite_approval_url(url: str, request_id: str) -> str:
-    normalized = url.replace("/approvals/", "/requests/")
-    prefix, _, _ = normalized.rpartition("/")
-    if prefix:
-        return f"{prefix}/{request_id}"
-    return request_id
+    return persist_approval_request(connection, request, now, oauth_source=oauth_source)
 
 
 def list_approval_requests(
