+++ src/codex_plugin_scanner/guard/store_connection_schema.py
@@ -40,6 +40,7 @@ from .store_storage_maintenance import (
     STORAGE_QUERY_INDEX_MIGRATION_VERSION,
     storage_maintenance_schema_statements,
 )
+from .store_watch_only_approval_schema import WATCH_ONLY_APPROVAL_MIGRATION_VERSION, ensure_watch_only_approval_schema
 from .store_workflow_capabilities_schema import (
     WORKFLOW_CAPABILITY_RECEIPT_EVENT_INDEX_MIGRATION_VERSION,
     ensure_workflow_capability_schema,
@@ -153,14 +154,13 @@ _POLICY_INDEX_STATEMENTS = (
 )
 
 _RECEIPT_WARN_ROLLUP_MIGRATION_VERSION = 16
-_WATCH_ONLY_APPROVAL_MIGRATION_VERSION = 23
 # Include the workflow-capability retired-index migration so a database created under an
 # earlier schema version (which still owns the retired index) is not treated as current and
 # is forced through ``_initialize_schema`` on the next open, where the index is reaped.
 _REQUIRED_SCHEMA_MIGRATION_VERSIONS = (
     *range(2, STORAGE_QUERY_INDEX_MIGRATION_VERSION + 1),
     WORKFLOW_CAPABILITY_RECEIPT_EVENT_INDEX_MIGRATION_VERSION,
-    _WATCH_ONLY_APPROVAL_MIGRATION_VERSION,
+    WATCH_ONLY_APPROVAL_MIGRATION_VERSION,
 )
 
 
@@ -1042,25 +1042,7 @@ class StoreConnectionSchemaMixin:
             self._ensure_approval_column(connection, "browser_intent_json", "text")
             self._ensure_approval_column(connection, "desktop_notified_at", "text")
             self._ensure_approval_column(connection, "raw_command_text", "text")
-            self._ensure_approval_column(connection, "guard_version", "text")
-            self._ensure_approval_column(connection, "first_seen_guard_version", "text")
-            self._ensure_approval_column(connection, "last_seen_guard_version", "text")
-            self._ensure_approval_column(connection, "oauth_source", "text")
-            self._ensure_approval_column(connection, "watch_only_observation", "integer not null default 0")
-            if not self._schema_version_applied(connection, version=_WATCH_ONLY_APPROVAL_MIGRATION_VERSION):
-                connection.execute(
-                    """
-                    update approval_requests
-                    set watch_only_observation = 1
-                    where coalesce(dedupe_count, 1) = 1
-                      and exists (
-                          select 1
-                          from json_each(coalesce(scanner_evidence_json, '[]'))
-                          where json_extract(value, '$.source') = 'observe_mode_inbox'
-                      )
-                    """
-                )
-                self._record_schema_version(connection, version=_WATCH_ONLY_APPROVAL_MIGRATION_VERSION)
+            ensure_watch_only_approval_schema(connection, schema=self)
             if not self._schema_version_applied(connection, version=3):
                 _backfill_approval_queue_columns_compat(connection)
                 self._record_schema_version(connection, version=3)
