+++ tests/fixtures/guard-command-corpus/decision-diff-report.json
@@ -12,7 +12,7 @@
     },
     "sources_sha256": {
       "source-0181b7136f7bc07d3738def8": "198f5e851641f0efc2187c46877c61d6049d9aac53943f205febbcf8ef29c45a",
-      "source-01febcca4015bcd1b7aeebc7": "716770a40fd0c295bb15c93aa43bcfd825d576a72f408559c781368d33e02481",
+      "source-01febcca4015bcd1b7aeebc7": "08e437fd34ddf3aeb0c330d74f39c75fea17aa296976f464afcb8b83e062cba6",
       "source-02697a55adfb35c263740ec9": "70763bb670c4a0afaa87f9929dc77a66086fd70c5d6611525209754b710e2ce3",
       "source-02aeab3ad023a11ada3b7743": "82eb33d49c84e6d2b907b0de245fc1ff9bc991af96124cff3071e9bb02fa1e34",
       "source-02c7cfdf8b65ef460436b48e": "5b5011a12fd773494eb3cfd9d56700a9f39a99ff8e9518deeeb29a7803271389",
@@ -29,6 +29,7 @@
       "source-0b67966321d400e77f4bdf0b": "bcc7127376cb25f6e2ae0992d2d4c81e43ad3cb920c0049fe66e85abd93981ee",
       "source-0b9ce3ba6ccc3dde2f1a2117": "323a1676d8e2c59f9d1aa484936be332111e9c56da50028749cb607c3aaf68de",
       "source-0bc9dc112696c30e81a95a4e": "31ae3f338de752bf5d122e9a8274184bc525c1fc4dbf4410f970bb380628e59c",
+      "source-0c09075e412cec66466763fb": "e05edc8973836637dcf40c8998ee29cbf2e7916fa33ce9a5f3713de9dd8cc82c",
       "source-0c31bcad3983f64fccbd8737": "3cb139c04f9a0076b25d5f76f280c0d0371740a9a00012b87592e21f67af162f",
       "source-0ce809a3ab9ff3e4e04be5b3": "8ecf88b1ad6b7b30eda2bbbd7a9329f47ddce856acb5d3bd3f8af13560e41384",
       "source-0cf0b6e67c56ec8bad1434fa": "66a1e556f79f9faec3ca095d5691585a8b0c075e10b626ed7a085755a99d10f5",
